First round of deleting ifdefs
