--- src/byterun/stdlib/delay.c
@@ -1,12 +1,0 @@
-void delay(int count){
-  #ifdef __AVR__
-  while(count--) {
-    _delay_ms(1);
-
-  }
-  #endif
-  #ifdef __PC__
-  printf("delay(%d)\n", count);
-  usleep((useconds_t) count * 1000);
-  #endif
-}
